app/backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.4
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
colorlog==6.9.0
courlan==1.3.2
dataclasses-json==0.6.7
dateparser==1.2.2
distro==1.9.0
docstring_parser==0.16
et_xmlfile==2.0.0
fastapi==0.116.1
fastembed==0.7.1
fastembed-haystack==1.5.0
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
haystack-ai @ git+https://github.com/deepset-ai/haystack.git@8ae24ca2687305a876c5949e0a8f9c1d1ba162d5
haystack-experimental==0.12.0
hf-xet==1.1.5
hpack==4.1.0
html2text==2025.4.15
htmldate==1.9.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
interchange==2021.0.4
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jq==1.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.1
jsonschema-specifications==2025.4.1
jusText==3.0.2
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.31
langchain-text-splitters==0.3.9
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.3
langsmith==0.4.16
lazy_imports==1.0.0
loguru==0.7.3
lxml==5.4.0
lxml_html_clean==0.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdit-plain==1.0.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
neo4j==5.28.2
networkx==3.5
numpy==2.3.1
onnxruntime==1.22.1
openai==1.101.0
openpyxl==3.1.5
orjson==3.11.2
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.1
pansi==2024.11.0
pillow==11.3.0
pluggy==1.6.0
portalocker==2.10.1
posthog==6.1.1
propcache==0.3.2
protobuf==6.31.1
py2neo==2021.2.4
py_rust_stemmers==0.1.5
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pypdf==5.8.0
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.14.3
qdrant-haystack==9.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
sentence-transformers==5.0.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
starlette==0.47.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tld==0.13.1
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.53.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.20.1
zstandard==0.24.0
